requirements.txt
numpy==1.15.4
scipy==1.1.0
pandas==0.23.4
scikit-learn==0.19.2
pandas-profiling==1.4.1
click==7.0
mlflow==0.7.0
apache-airflow[gcp_api]==1.10.0
